Allow re-execution of the %prep section
With Yocto, when building rpm a second time, the do_prep() function
fails because it assumes it hasn't already ran but will be re-using the
previous build directory which hasn't been cleaned. This patch prevent
such a situation to fail.

This issue is not present when building with GBS since every new build
is made in a clean directory but since the yocto recipes are generated
from spec file, it is better to directly correct the spec file.
This patch has no impact on the GBS build.

Bug-Tizen: BTY-15
Change-Id: I1e992b63b41ae76a4d6706046e741dc2d25b15eb
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>